Refactor assertCreateTableAsSelect() and assertExplainAnalyze() functions
